+++ drivers/net/usb/cx82310_eth.c
@@ -352,7 +352,7 @@ static const struct driver_info	cx82310_info = {
 static const struct usb_device_id products[] = {
 	{
 		USB_DEVICE_CLASS(0x0572, 0xcb01, 0xff, 0, 0),
-		.driver_info = (unsigned long) &cx82310_info
+		.driver_info = (uintptr_t) &cx82310_info
 	},
 	{ },
 };
