+++ drivers/net/usb/sr9700.c
@@ -530,7 +530,7 @@ static const struct driver_info sr9700_driver_info = {
 static const struct usb_device_id products[] = {
 	{
 		USB_DEVICE(0x0fe6, 0x9700),	/* SR9700 device */
-		.driver_info = (unsigned long)&sr9700_driver_info,
+		.driver_info = (uintptr_t)&sr9700_driver_info,
 	},
 	{},			/* END */
 };
